+++ inc/resources/ReferableFactory.h
@@ -80,7 +80,7 @@ public:
 	/**
 	\return The total number of types
 	*/
-	virtual u32 GetTypeCount() const = 0;
+	virtual size_t GetTypeCount() const = 0;
 
 	//! Creates a new unused id.
 	virtual lxID MakeId() = 0;
